code cleanups and bugs fixes
